pyproject.toml
[project]
name = "ticketing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "debugpy>=1.8.19",
    "dj-rest-auth>=5.0.2",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.10",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]

[tool.django-stubs]
django_settings_module = "backend.config.settings"

[dependency-groups]
dev = [
    "debugpy>=1.8.19",
    "django-stubs>=5.2.8",
    "djangorestframework-stubs>=3.16.6",
    "ty>=0.0.8",
]
